Archivos de postproceso vtk escritos en lenguaje binario

+++ OpenACC/SIMPLE2D/mod_malla.f90
@@ -8,9 +8,9 @@
 !
 MODULE malla
   implicit none
-  INTEGER, PARAMETER :: mi=901, nj=501, nsolid = 7
-  INTEGER, PARAMETER :: SGL=SELECTED_REAL_KIND(P=6,R=37)
-  INTEGER, PARAMETER :: DBL=SELECTED_REAL_KIND(P=15,R=307)
+  INTEGER(4), PARAMETER :: mi=901, nj=501, nsolid = 7
+  INTEGER(4), PARAMETER :: SGL=SELECTED_REAL_KIND(P=6,R=37)
+  INTEGER(4), PARAMETER :: DBL=SELECTED_REAL_KIND(P=15,R=307)
   character(len=9), parameter :: form24="(3D23.15)",form25="(4D23.15)"
   character(len=9), parameter :: form26="(6D23.15)",form27="(10e15.7)"
   REAL(kind=DBL),   PARAMETER :: cero=0._DBL, visc_solido = 1.e40_DBL
